src/requirements.txt
# requirements
numpy == 1.23.1
pyusb == 1.2.1

# dev
black == 22.6.0
tomli == 2.0.1

# test
